refactor: maxLength 제한 로직 훅으로 분리
